test iniciados

+++ src/Library/Attack.cs
@@ -7,7 +7,7 @@ public class Attack : IAttack
     public double Accuracy {get; private set;}
     public int Power {get; private set;}
 
-    public Attack(string name, Type type, int Accuracy, int Power)
+    public Attack(string name, Type type, double Accuracy, int Power)
     {
         this.Name = name;
         this.Type = type;
